// src/types/FormValueInitialState.ts
export type FormValueInitialState = {
  guarantyNumber: string;
  proofOfPurchase: string;
  name: string;
  phone: string | number;
  email: string;
  town: string;
  street: string;
  productName: string;
  SN: string;
  buyingDate: string;
  faultDate: string;
  faultInfo: string;
  deviceStatus: string;
  customerRequest: string;
};